services/api_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
PyJWT==2.10.1
asyncpg==0.29.0
httpx==0.25.0
paramiko==3.4.0
robotframework==7.0
robotframework-sshlibrary==3.8.0